Fixes third argument in call to output->write
The third parameter of `Symfony\Component\Console\Output\OutputInterface::write()` is `$type`; not `$verbosity`. This is intended for distinguishing between normal, raw, and plain output. 

The previous code happens to work in most cases because these constants are similar.

```
interface OutputInterface
{
    const VERBOSITY_QUIET = 0;
    const VERBOSITY_NORMAL = 1;
    const VERBOSITY_VERBOSE = 2;
    const VERBOSITY_VERY_VERBOSE = 3;
    const VERBOSITY_DEBUG = 4;

    const OUTPUT_NORMAL = 0;
    const OUTPUT_RAW = 1;
    const OUTPUT_PLAIN = 2;
```

When VERBOSITY_DEBUG is used, an InvalidArgumentException is thrown with the message `Unknown output type given (4)`.